+++ src/Configuration/SwiftMailer/MessageConfigurationInterface.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * TechDivision\Import\Callbacks\CallbackInterface
+ * TechDivision\Import\Configuration\SwiftMailer\MessageConfigurationInterface
  *
  * NOTICE OF LICENSE
  *
@@ -18,10 +18,10 @@
  * @link      http://www.techdivision.com
  */
 
-namespace TechDivision\Import\Callbacks;
+namespace TechDivision\Import\Configuration\SwiftMailer;
 
 /**
- * interface for all callback implementations.
+ * The swift mailer message configuration interface.
  *
  * @author    Tim Wagner <t.wagner@techdivision.com>
  * @copyright 2016 TechDivision GmbH <info@techdivision.com>
@@ -29,16 +29,13 @@ namespace TechDivision\Import\Callbacks;
  * @link      https://github.com/techdivision/import
  * @link      http://www.techdivision.com
  */
-interface CallbackInterface
+interface MessageConfigurationInterface
 {
 
     /**
-     * Will be invoked by a observer it has been registered for.
+     * Return's the swift mailer message factory used to create he instance.
      *
-     * @param string $attributeCode  The code of the attribute the passed value is for
-     * @param mixed  $attributeValue The value to handle
-     *
-     * @return mixed The modified value
+     * @return string The swift mailer message factory
      */
-    public function handle($attributeCode, $attributeValue);
+    public function getMessageFactory();
 }